backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-extensions
django-filter
djangorestframework-simplejwt
python-decouple
psycopg2-binary
Pillow
PyMuPDF
PyPDF2
pdfplumber
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pytesseract
spacy
nltk
scikit-learn
pandas
numpy
openpyxl
requests
gunicorn
whitenoise
celery
redis
python-dateutil
pytz
tzdata
torch
transformers
joblib
